[1,2,3 단계] auth 기능 구현 (#7)
* [1단계] refactor: 패스워드를 암호화하도록 리팩토링

* [1단계] feat: 로그인 및 엑세스 토큰 발급 구현

- subject는 user ID
- role은 미구현
- auth를 가장 상위 모듈로 두고, 의존성이 섞이지 않도록 주의할것
  - member -> auth (O)
  - auth -> member (X)

* [1단계] feat: 자기정보 조회 구현

* [1단계] chore: 잘 못된 displayName 수정

* [1단계] feat: CORS를 모조리 열어줘버리는 Free한 mind의 소유자

* [2단계] feat: 토큰을 통한 권한을 확인하는 기능 추가

* [2단계] feat: 테스트 및 기능 확인을 위한 마스터 토큰 추가

* [2단계] feat: 예약 및 예약 취소에 권한 확인 추가

* [2단계] feat: 예약이 Member를 참조하도록 수정

* [3단계] feat: 권한체계 추가

* [3단계] feat: 어드민 계정 무적박살맨 삽입

- CommandLineRunner 이용

* [3단계] feat: 어드민 계정 삭제기능 추가

* [3단계] feat: '/admin/**' path에 대한 권한 제어

+++ src/main/java/nextstep/reservation/ReservationService.java
@@ -1,5 +1,6 @@
 package nextstep.reservation;
 
+import nextstep.member.Member;
 import nextstep.schedule.Schedule;
 import nextstep.schedule.ScheduleDao;
 import nextstep.support.DuplicateEntityException;
@@ -21,7 +22,7 @@ public class ReservationService {
         this.scheduleDao = scheduleDao;
     }
 
-    public Long create(ReservationRequest reservationRequest) {
+    public Long create(ReservationRequest reservationRequest, Member member) {
         Schedule schedule = scheduleDao.findById(reservationRequest.getScheduleId());
         if (schedule == null) {
             throw new NullPointerException();
@@ -34,7 +35,7 @@ public class ReservationService {
 
         Reservation newReservation = new Reservation(
                 schedule,
-                reservationRequest.getName()
+                member
         );
 
         return reservationDao.save(newReservation);
@@ -49,12 +50,21 @@ public class ReservationService {
         return reservationDao.findAllByThemeIdAndDate(themeId, date);
     }
 
-    public void deleteById(Long id) {
+    public void deleteById(Long id, Member member) {
         Reservation reservation = reservationDao.findById(id);
+
         if (reservation == null) {
             throw new NullPointerException();
         }
 
+        if (NotAdminAndNotMine(member, reservation)) {
+            throw new IllegalStateException("내 예약만 삭제할 수 있습니다.");
+        }
+
         reservationDao.deleteById(id);
+    }
+
+    private boolean NotAdminAndNotMine(Member member, Reservation reservation) {
+        return !reservation.getMember().equals(member) && !member.isAdmin();
     }
 }
